requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
openai==0.28.1
python-dotenv==1.0.0
gunicorn==21.2.0